# aio-database-api/openapi.yaml
openapi: 3.1.0
info:
   title: Copped AIO database API
   description: Documentation for the [Copped AIO](https://aio.copped-inc.com) database API
   version: 0.0.1
servers:
   - url: https://database.copped-inc.com
...
paths:
...
   /captcha/preharvest:
      get:
         tags:
            - captcha
         summary: Get captcha preharvest tasks
         description: Retrieve all captcha preharvest tasks
         operationId: getPreharvestTasks
         parameters:
            - $ref: "#/components/parameters/sort"
            - $ref: "#/components/parameters/limit"
            - $ref: "#/components/parameters/after"
            - $ref: "#/components/parameters/before"
         responses:
            "200":
               description: An array of preharvest task(s).
               content:
                  application/json:
                     schema:
                        type: array
                        items:
                           $ref: "#/components/schemas/preharvestTask"
                     examples:
                        preharvestTasks:
                           summary: Captcha preharvest tasks
                           description: A list of captcha preharvest tasks
            "204":
               description: No captcha preharvest tasks found
            "401":
               $ref: "#/components/responses/401"
            "500":
               $ref: "#/components/responses/500"
            "5XX":
               $ref: "#/components/responses/5XX"
         security:
            - admin: []
   /captcha/preharvest/{task-id}:
      patch:
         tags:
            - captcha
         summary: Edit captcha preharvest task
         description: Edit the specified captcha preharvest task
         operationId: editPreharvestTask
         parameters:
            - description: The captcha preharvest task to be edited
              $ref: "#/components/parameters/task-id"
         requestBody:
            content:
               application/json:
                  schema:
                     type: object
                     description:
                     minProperties: 1
                     properties:
                        state:
                           $ref: "#/components/schemas/preharvestTaskState"
                        uses:
                           type: integer
                           minimum: 0
                           maximum: 7
                        schedule:
                           $ref: "#/components/schemas/preharvestTaskSchedule"
                        date:
                           $ref: "#/components/schemas/date"
                  examples:
                     updateUses:
                        summary: Update uses
                        description: Update the remaining uses of the captcha preharvest task.
                        value:
                           uses: 2
                     updateSchedule:
                        summary: Update schedule
                        description: Update the captcha preharvest task's schedule.
                        value:
                           schedule: every 2 days
            required: true
         responses:
            "200":
               description: Captcha preharvest task edited successfully
               content:
                  application/json:
                     schema:
                        $ref: "#/components/schemas/notification"
                     examples:
                        globalNotification:
                           summary: Global notification
                           description: The edited global notification.
                           value:
                              id: "rRij2j86JzU8rpQP9IBx"
                              created_at: "2023-08-30T11:33:47.742543Z"
                              title: "foo"
                              text: "baz"
            "304":
               $ref: "#/components/responses/304"
            "400":
               $ref: "#/components/responses/400"
            "401":
               $ref: "#/components/responses/401"
            "404":
               description: "Captcha preharvest task not found"
               $ref: "#/components/responses/404"
            "500":
               $ref: "#/components/responses/500"
            "5XX":
               $ref: "#/components/responses/5XX"
         security:
            - admin: []
...
components:
   parameters:
...
      task-id:
         name: task-id
         in: path
         schema:
            type: string
         required: true
      sort:
         name: sort
         description: |
            Sorting order options provided by firestore:  
            * **1** ascending
            * **2** descending
         in: query
         schema:
            type: integer
            format: int32
            default: 2
            enum:
               - 1
               - 2
      limit:
         description: Limit the number of results for endpoints that return an array of items.
         name: limit
         in: query
         schema:
            type: integer
            minimum: 1
      after:
         description: Used for pagination. Provide a unix timestamp value to return results after / before the date specified (depending on the **sorting order** [ascending / descending]).
         name: after
         in: query
         schema:
            type: integer
            minimum: 0
            format: unix timestamp
      before:
         name: before
         description: |
            Used for pagination. Provide a unix timestamp value to return results after / before the date specified (depending on the **sorting order** [descending / ascending]).  
            Must be after / before the date specified in the `after` query parameter (depending on the **sorting order** [ascending / descending]).
         in: query
         schema:
            type: integer
            minimum: 0
            format: unix timestamp
   responses:
...
      "304":
         description: Not modified
         content:
            text/plain:
               schema:
                  type: string
               examples:
                  notModified:
                     summary: Not modified
                     value: Not modified
      "400":
         description: Bad request
         content:
            text/plain:
               schema:
                  type: string
               examples:
                  badRequest:
                     summary: Bad request
                     value: Bad request
      "401":
         description: Request unauthorized
         content:
            text/plain:
               schema:
                  type: string
               examples:
                  unauthorized:
                     summary: Unauthorized
                     value: Unauthorized
...
      "404":
         description: Not found
         content:
            text/plain:
               schema:
                  type: string
               examples:
                  internalServerError:
                     summary: Not found
                     value: Not found

            text/html:
               schema:
                  description: Redirect to https://aio.copped-inc.com/utility/404
      "500":
         description: Internal server error
         content:
            text/plain:
               schema:
                  type: string
               examples:
                  internalServerError:
                     summary: Internal server error
                     value: Internal server error
      "5XX":
         description: Unexpected/-handled error
         content:
            text/plain:
               schema:
                  type: string
            text/html; charset=UTF-8:
               schema:
                  type: string
...
   schemas:
      notification:
         description: |
            #### Notifications  
            Notifications are used to convey both personal and public messages and information to users. Thus, they can be grouped into **global** and "normal" notifications, referred to as **user** notifications to avoid ambiguities.  
            Global and user notifications use distinct endpoints and are handled different internally. Nonetheless, they share the same schema utilizing different fields and are built upon the same infrastructure.
            ### Global notifications
            Global notifications get shared with all users.  
            To allow for editing content of a global notification afterwards, users don't receive simple copies of global notifications. Instead, their user notification shares its ID with the global notification and has `global` set to `true`, indicating it's referencing a global notification.
            ```
            In global notifications, the global and read fields aren't present.
            ```
            ### User notifications
            User notifications (simply notifications) are stored individually for each user. They consist of both personal and shared notifications.  
            Shared notifications (`global: true`) are a special case, as they reference a global notification. While the shared notification doesn't include the `text` nor `title` fields, it has the same ID as the global notification it's referencing.  
            To get the actual shared message's content, a subsequent call to retrieve the global notification's content is required for each shared notification.
            ```
            Shared notifications don't include the title and text fields. Additionally, user notifications always include the read field.
            ```
         allOf:
            - type: object
              properties:
                 id:
                    type: string
                    minLength: 1
                 created_at:
                    $ref: "#/components/schemas/date"
                 global:
                    type: boolean
                    default: false
              required:
                 - id
                 - created_at
            - $ref: "#/components/schemas/notificationEdit"
...
      notificationEdit:
         description: "`read` field will be ignored when editing global messages"
         type: object
         properties:
            title:
               type: string
               minLength: 1
            text:
               type: string
               minLength: 1
            read:
               type: boolean
               default: false
...
      preharvestTask:
         description: |
            #### Captcha preharvest tasks
            ... are used to generate captchas at certain times. This aims to pre-generate solved captchas to allow for more checkouts by having captchas readily available instead of having to wait for them.
         allOf:
            - type: object
              description:
              properties:
                 id:
                    type: string
                    minLength: 1
                 user_id:
                    type: string
                    minLength: 1
                 state:
                    $ref: "#/components/schemas/preharvestTaskState"
                 uses:
                    type: integer
                    default: 0
                    minimum: 0
              required:
                 - id
                 - user_id
                 - state
            - $ref: "#/components/schemas/preharvestTaskCreate"
      preharvestTaskCreate:
         type: object
         properties:
            site:
               $ref: "#/components/schemas/site"
            routine:
               type: boolean
               default: false
            date:
               $ref: "#/components/schemas/date"
            schedule:
               $ref: "#/components/schemas/preharvestTaskSchedule"
         required:
            - site
      preharvestTaskSchedule:
         type: string
         pattern: ^(?si:(?:every (?:(2|3|4|5|6) (day|week|month)(?:s)?|(day|week|month)))|(daily|weekly|monthly))$
      preharvestTaskState:
         type: integer
         description: |
            The current state of a captcha preharvest task:
            * **1** running
            * **2** stopped
         minimum: 1
         maximum: 2
      date:
         type: string
         minLength: 1
         format: date-time
      site:
         type: string
         description: |
            #### Sites
            Sites represent the modules and functionality of the Copped AIO bot. Every site has further metadata storing those properties.
            Currently, the list of sites is as following:
            * **kith_eu** - [KithEU](https://eu.kith.com)
            * **queue_it** - [Queue-it](https://queue-it.com)
         enum:
            - kith_eu
            - queue_it
         externalDocumentation:
            description: Source defition of the Site type.
            url: https://github.com/Copped-Inc/aio-types/blob/v2/modules/site.go
   securitySchemes:
      admin:
         description: Authorization for internal traffic and certain endpoints
         type: apiKey
         in: header
         name: Password
      user:
         description: Authorization used by normal users and thus by most generic request
         type: apiKey
         in: cookie
         name: authorization
...